requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.9
aiosignal==1.3.1
amqp==5.2.0
annotated-types==0.7.0
anyio==4.6.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.8.1
asttokens==2.4.1
async-lru==2.0.4
attrs==24.2.0
babel==2.16.0
beautifulsoup4==4.12.3
billiard==4.2.1
bleach==6.1.0
blinker==1.8.2
cachelib==0.13.0
celery==5.3.4
shortuuid
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
comm==0.2.2
contourpy==1.3.0
crispy-bootstrap5==2024.2
crispy-bootstrap4==2023.1
crispy-tailwind==1.0.3
cycler==0.12.1
debugpy==1.8.6
decorator==5.1.1
defusedxml==0.7.1
distro==1.9.0
Django==4.2.9
django-api-generator==1.0.17
django-celery-results==2.5.1
django-cors-headers==4.4.0
django-crispy-forms==2.3
django-debug-toolbar==4.2.0
#django-fullcalendar
djangorestframework==3.14.0
drf-spectacular==0.26.5
executing==2.1.0
fastjsonschema==2.20.0
filelock==3.16.1
Flask==3.0.3
Flask-Session==0.8.0
fonttools==4.54.1
fqdn==1.5.1
frozenlist==1.4.1
fsspec==2024.9.0
gunicorn==21.2.0
h11==0.14.0
hiredis==2.2.3
httpcore==1.0.5
httpx==0.27.2
huggingface-hub==0.25.2
HyperPyYAML==1.2.2
idna==3.10
inflection==0.5.1
ipykernel==6.29.5
ipython==8.27.0
ipywidgets==8.1.5
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
jiter==0.5.0
joblib==1.4.2
json5==0.9.25
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.2.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
kiwisolver==1.4.7
kombu==5.4.2
MarkupSafe==2.1.5
matplotlib==3.9.2
matplotlib-inline==0.1.7
mistune==3.0.2
mpmath==1.3.0
msgspec==0.18.6
multidict==6.1.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.3
notebook==7.2.2
notebook_shim==0.2.4
numpy==2.1.1
openai==0.28.0
overrides==7.7.0
packaging==24.1
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
Pillow==10.0.1
pipdeptree==2.23.4
platformdirs==4.3.6
prometheus_client==0.21.0
prompt_toolkit==3.0.47
psutil==6.0.0
pure_eval==0.2.3
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-json-logger==2.0.7
pytz==2024.2
pywinpty==2.0.13; platform_system == "Windows"
PyYAML==6.0.2
pyzmq==26.2.0
redis==4.5.4
referencing==0.35.1
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.20.0
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
scikit-learn==1.5.2
scipy==1.14.1
seaborn==0.13.2
Send2Trash==1.8.3
sentencepiece==0.2.0
setuptools==75.1.0
six==1.16.0
sniffio==1.3.1
soundfile==0.12.1
soupsieve==2.6
speechbrain==1.0.1
sqlparse==0.5.1
stack-data==0.6.3
str2bool==1.1
sympy==1.13.3
terminado==0.18.1
threadpoolctl==3.5.0
tinycss2==1.3.0
torch==2.2.2
torchaudio==2.2.2
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
types-python-dateutil==2.9.0.20240906
typing_extensions==4.12.2
tzdata==2024.2
uri-template==1.3.0
uritemplate==4.1.1
urllib3==2.2.3
vine==5.1.0
wcwidth==0.2.13
webcolors==24.8.0
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.0.4
whitenoise==6.5.0
widgetsnbextension==4.0.13
yarl==1.13.1
django-widget-tweaks
google-api-python-client==2.166.0